requirements.txt
requests==2.31.0
osu.py==1.1.0
